KUDU-3112. Fix WaitForBind method for checking service status
The WaitForBind method is used to check a service status according to
outputs of 'lsof' command, in this method, it will wait for the service
listening at a port number. If a the outputs of 'lsof' command includes
'n*:<port-num>', it means the service listen all the addresses on the
port number, but when we pass a specific address to the method,
e.g. 127.0.0.1, and the 'lsof' command show 'n*:<port-num>', this
also raise error.

Change-Id: Ib79297e0eb59cc96a91c6e301f6a70ba123f4644